test: stabilize matheron sampling checks
Agent-Logs-Url: https://github.com/jejjohnson/gaussx/sessions/94f33a48-1a6f-45d0-851b-f33a94cafadc

+++ src/gaussx/_gp/_matheron.py
@@ -46,7 +46,7 @@ def matheron_update(
     """
     prior_sample_target = jnp.asarray(prior_sample_target)
     prior_sample_conditioning = jnp.asarray(prior_sample_conditioning)
-    observed_value = jnp.asarray(observed_value, dtype=prior_sample_conditioning.dtype)
+    observed_value = jnp.asarray(observed_value)
 
     if prior_sample_target.ndim != 2:
         raise ValueError("prior_sample_target must have shape (S, N_star).")
